fix OffscreenCanvas getContext warn about willReadFrequently
